[PD] use int32 for kv indices & get num_reserved_decode_tokens from server_args (#7214)

+++ python/sglang/srt/disaggregation/fake/conn.py
@@ -48,7 +48,7 @@ class FakeKVSender(BaseKVSender):
 
     def send(
         self,
-        kv_indices: npt.NDArray[np.int64],
+        kv_indices: npt.NDArray[np.int32],
     ):
         self.has_sent = True
         logger.info(f"FakeKVSender send with kv_indices: {kv_indices}")
